SAPP-14 - Delete project allowed only to the project admin

+++ tasks-app/src/main/java/com/tasks/business/ProjectService.java
@@ -6,6 +6,7 @@ import com.tasks.business.exceptions.DuplicatedResourceException;
 import com.tasks.business.exceptions.InstanceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PostAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import com.tasks.business.repository.ProjectsRepository;
 import com.tasks.business.repository.UsersRepository;
@@ -65,12 +66,16 @@ public class ProjectService {
     }
     
     @Transactional()
-    public void removeById(Long id) throws InstanceNotFoundException {
+    public boolean removeById(Long id) throws InstanceNotFoundException {
         Optional<Project> optTask = projectsRepository.findById(id);
         if(!optTask.isPresent()) {
             throw new InstanceNotFoundException(id, "Project" , MessageFormat.format("Project {0} does not exist", id));
         }
-        projectsRepository.delete(optTask.get());
+        if(optTask.get().getAdmin().getUsername().compareTo(SecurityContextHolder.getContext().getAuthentication().getName()) == 0 ) {
+            projectsRepository.delete(optTask.get());
+            return true;
+        }
+        return false;
     }
     
     @Transactional(readOnly = true)
